Fix very poor compile-time in PR19499 due to excessive tree walks in domtree. When finding a nearest common dominator, if neither A dominates B nor B dominates A, we immediately resorted to a tree walk. The tree walk here is *particularly* expensive because we have to build a (potentially very large) set for one side's dominators and compare it with the other side's.
If at any point we have DFS info, we don't need to do any of this. We
can just walk up one side's immediate dominators and return the first
one which dominates the other side. Because of the DFS info, the
dominates queries are trivially constant time.

This reduces the optimizers time in the test case on PR19499 by 70%. It
now optimizes in about 30 seconds for me. And there is still more to be
done for this case.

git-svn-id: https://llvm.org/svn/llvm-project/llvm/trunk@207406 91177308-0d34-0410-b5e6-96231b3b80d8