Merge r245577 into branch_37
[ARM] Don't try and custom lower a vNi64 SETCC.

It won't go well. We've already marked 64-bit SETCCs as non-Custom, but it's
just possible that a SETCC has a legal result type but an illegal operand
type. If this happens, bail out before we create unselectable nodes.

Fixes PR24292. I tried to create a testcase but in 99% of cases we can't
trigger this - not surprising that this bug has been latent since 2009.


git-svn-id: https://llvm.org/svn/llvm-project/llvm/branches/release_37@245578 91177308-0d34-0410-b5e6-96231b3b80d8