Add support for importing and exporting Registry objects on Windows
On Windows, the DLL containing the registry will get its own global head
and tail variables, so the entries registered in the DLL will be
invisible to the consumer.

In order to solve this, we need to export a getter function from the
plugin DLL per registry and copy over the data inside it.  This patch
adds support for this.  This will be used to support clang plugins on
Windows.

git-svn-id: https://llvm.org/svn/llvm-project/llvm/trunk@260261 91177308-0d34-0410-b5e6-96231b3b80d8