[MIR] Teach the mir printer how to print the register bank.
For now, we put the register bank in the Class field since a register
may only have one of those at a given time. The downside of that
representation is that if a register class and a register bank have the
same name, we will not be able to distinguish them.

git-svn-id: https://llvm.org/svn/llvm-project/llvm/trunk@265796 91177308-0d34-0410-b5e6-96231b3b80d8