test: Be consistent with clang's sanitizer lit config
The logic that sets up lit features for sanitizers is largely copied
between here and clang, except clang's was fixed some time ago to
handle multiple sanitizers (ie, Asan + Ubsan). This just makes the
code in LLVM consistent with how it's done in clang to avoid any
gotchas by users of this.

git-svn-id: https://llvm.org/svn/llvm-project/llvm/trunk@270510 91177308-0d34-0410-b5e6-96231b3b80d8