Fix gather-root.ll SLP vectorizer test to not expose UB.
The undefined behaviour (signed integer overflow) is not a regression
in itself as it was already there, but the test exposing it is a
regression compared to rc1, i.e. the lit tests no longer run ubsan-clean.

This commit fixes the test based on Matt's change in r279125 to not
expose the undefined behaviour.


git-svn-id: https://llvm.org/svn/llvm-project/llvm/branches/release_39@279468 91177308-0d34-0410-b5e6-96231b3b80d8