[AVX-512] Fix masked VPERMI2PS isel when the index comes from a bitcast.
We need to bitcast the index operand to a floating point type so that it matches the result type. If not then the passthru part of the DAG will be a bitcast from the index's original type to the destination type. This makes it very difficult to match. The other option would be to add 5 sets of patterns for every other possible type.

git-svn-id: https://llvm.org/svn/llvm-project/llvm/trunk@280696 91177308-0d34-0410-b5e6-96231b3b80d8