[CVP] Adjust iteration order to reduce the amount of work required
CVP doesn't care about the order of blocks visited, but by using a pre-order traversal over the graph we can a) not visit unreachable blocks and b) optimize as we go so that analysis of later blocks produce slightly more precise results.

I noticed this via inspection and don't have a concrete example which points to the issue.  



git-svn-id: https://llvm.org/svn/llvm-project/llvm/trunk@290760 91177308-0d34-0410-b5e6-96231b3b80d8