[ARM] Add curly braces around switch case [NFC]
My previous commit r304702 introduced a new case into a switch statement.
This case defined a variable but I forgot to add the curly brackets around the
case to limit the scope.

This change puts the curly braces back in so that the next person that adds a
case doesn't get a build failure. Thanks to avieira for the spot.

Differential Revision: https://reviews.llvm.org/D33931



git-svn-id: https://llvm.org/svn/llvm-project/llvm/trunk@304785 91177308-0d34-0410-b5e6-96231b3b80d8