[x86][SSE] Add widenSubVector helper. NFCI.
Helper function to insert a subvector into the bottom elements of a larger zero/undef vector with the same scalar type.

I've converted a couple of INSERT_SUBVECTOR calls to use it, there are plenty more although in some cases I was worried it might make the code more ambiguous. 

git-svn-id: https://llvm.org/svn/llvm-project/llvm/trunk@327236 91177308-0d34-0410-b5e6-96231b3b80d8