[MemorySSA] Turn an assert into a condition
Eli pointed out that variadic functions are totally a thing, so this
assert is incorrect.

No test-case is provided, since the only way this assert fires is if a
specific DenseMap falls back to doing `isEqual` checks, and that seems
fairly brittle (and requires a pyramid of growing
`call void (i8, ...) @varargs(i8 0)`).


git-svn-id: https://llvm.org/svn/llvm-project/llvm/trunk@328755 91177308-0d34-0410-b5e6-96231b3b80d8