[dsymutil] Apply recursion workaround for threading
The DwarfLinker can have some very deep recursion that can max out the
(significantly smaller) stack when using threads. We don't want this
limitation when we only have a single thread. We already have this
workaround for the architecture-related threading. This patch applies
the same workaround to the parallel analysis and cloning.

Differential revision: https://reviews.llvm.org/D45172

git-svn-id: https://llvm.org/svn/llvm-project/llvm/trunk@329093 91177308-0d34-0410-b5e6-96231b3b80d8