[Loopinfo] Remove one latch-case in getLoopID. NFC.
getLoopID has different control flow for two cases: If there is a
single loop latch and for any other number of loop latches (0 and more
than one). The latter case should return the same result if there is
only a single latch. We can save the preceding redundant search for a
latch by handling both cases with the same code.

Differential Revision: https://reviews.llvm.org/D52118

git-svn-id: https://llvm.org/svn/llvm-project/llvm/trunk@342406 91177308-0d34-0410-b5e6-96231b3b80d8