[BPF][BTF] Add BTF generation if the target is BPF
BTF is the debug format for BPF, which contains
two parts:
  .BTF contains all the type and string information, and
  .BTF.ext contains the func_info and line_info.

The type and func_info are gathered during CodeGen/AsmPrinter
where dwarf debug_info is generated, and line_info is
gathered in MCObjectStreamer before writing to
the object file.

After all the information is gathered, the two
sections are emitted in MCObjectStreamer.

Signed-off-by: Song Liu <songliubraving@fb.com>
Signed-off-by: Yonghong Song <yhs@fb.com>