[x86] add tests for select-of-fp-constants; NFC
There are many options here depending on subtarget,
but we are uniformly relying on a transform that was 
driven by performance for a 32-bit SSE2 target in 2009.

Note: The same motivation was apparently used to do this 
transform for *all* targets, so non-x86 may want to look
at this too.



git-svn-id: https://llvm.org/svn/llvm-project/llvm/trunk@347525 91177308-0d34-0410-b5e6-96231b3b80d8