AMDGPU: Record SGPR spills when restoring too
It's possible in some cases to have a restore present
without a corresponding spill. Due to an apparent bug
in D54366 <https://reviews.llvm.org/D54366>, only the
restore for a register was emitted. It's probably
always a bug for this to happen, but due to how SGPR
spilling is implemented, this makes the issues appear
worse than it is.

git-svn-id: https://llvm.org/svn/llvm-project/llvm/trunk@347595 91177308-0d34-0410-b5e6-96231b3b80d8