DebugInfo: Fix for missing comp_dir handling with r349207
When deciding lazily whether a CU would be split or non-split I
accidentally dropped some handling for the line tables comp_dir (by
doing it lazily it was too late to be handled properly by the MC line
table code).

Move that bit of the code back to the non-lazy place.

git-svn-id: https://llvm.org/svn/llvm-project/llvm/trunk@349819 91177308-0d34-0410-b5e6-96231b3b80d8