[gold] Fix test after BitStream reader error changes
The recent change to the BitStream reader error handling in r364464
changed the error message format (from "LLVM ERROR:" to just "error"),
leading to a failure in this test which is only executed for very recent
versions of gold. Fix this by removing that part of the error message
check, leaving only the interesting part of the message to be checked.

git-svn-id: https://llvm.org/svn/llvm-project/llvm/trunk@364965 91177308-0d34-0410-b5e6-96231b3b80d8