[ARM] MVE floating point compares and selects
Much like integers, this adds MVE floating point compares and select. It
requires a lot more buildvector/shuffle code because we may need to expand the
compares without mve.fp, and requires support for and/or because of the way we
lower llvm condition codes.

Some original code by David Sherwood

Differential Revision: https://reviews.llvm.org/D65054


git-svn-id: https://llvm.org/svn/llvm-project/llvm/trunk@366909 91177308-0d34-0410-b5e6-96231b3b80d8